refactor: Change the package structure to src-layout
