+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "emadam-password-generator"
+name = "rpgen"
 version = "0.0.1"
 authors = [
   { name="Emadam", email="ldy9037@naver.com" },
@@ -19,3 +19,8 @@ classifiers = [
 
 [project.urls]
 "Homepage" = "https://github.com/ldy9037"
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "src"
+]
