+++ requirements.txt
@@ -1,0 +1,44 @@
+attrs==22.2.0
+black==23.1.0
+bleach==6.0.0
+build==0.10.0
+certifi==2022.12.7
+charset-normalizer==3.1.0
+click==8.1.3
+docutils==0.19
+-e git+ssh://git@github.com-personal/ldy9037/python-password-generator.git@5471b744ac0ee1bc5181d879ae093d459acd80eb#egg=emadam_password_generator
+exceptiongroup==1.1.0
+flake8==6.0.0
+idna==3.4
+importlib-metadata==6.0.0
+iniconfig==2.0.0
+jaraco.classes==3.2.3
+keyring==23.13.1
+markdown-it-py==2.2.0
+mccabe==0.7.0
+mdurl==0.1.2
+more-itertools==9.1.0
+mypy==1.0.1
+mypy-extensions==1.0.0
+packaging==23.0
+pathspec==0.11.0
+pkginfo==1.9.6
+platformdirs==3.1.0
+pluggy==1.0.0
+pycodestyle==2.10.0
+pyflakes==3.0.1
+Pygments==2.14.0
+pyproject_hooks==1.0.0
+pytest==7.2.2
+readme-renderer==37.3
+requests==2.28.2
+requests-toolbelt==0.10.1
+rfc3986==2.0.0
+rich==13.3.2
+six==1.16.0
+tomli==2.0.1
+twine==4.0.2
+typing_extensions==4.5.0
+urllib3==1.26.15
+webencodings==0.5.1
+zipp==3.15.0
